+++ Gruntfile.js
@@ -25,7 +25,6 @@ module.exports = function(grunt) {
     dist_firefox: 'dist/firefox/',
 
     src: 'src/all/',
-    test: 'test/',
     src_addon: 'src/all/background_page/',
     src_addon_vendors: 'src/all/background_page/vendors/',
     src_chrome: 'src/chrome/',
@@ -60,19 +59,17 @@ module.exports = function(grunt) {
   grunt.registerTask('bundle-firefox', ['copy:manifest_firefox', 'bundle', 'browserify:vendors']);
   grunt.registerTask('bundle-chrome', ['copy:manifest_chrome', 'bundle', 'browserify:vendors']);
 
-  grunt.registerTask('bundle-test', ['browserify:test', 'copy:test']);
-
   grunt.registerTask('build', ['build-firefox', 'build-chrome']);
 
   grunt.registerTask('build-firefox', ['build-firefox-debug', 'build-firefox-prod']);
-  grunt.registerTask('build-firefox-debug', ['clean:build', 'pre-dist', 'copy:config_debug', 'bundle-firefox', 'bundle-test', 'shell:build_quickaccess_debug', 'shell:build_firefox_debug']);
+  grunt.registerTask('build-firefox-debug', ['clean:build', 'pre-dist', 'copy:config_debug', 'bundle-firefox', 'shell:build_quickaccess_debug', 'shell:build_firefox_debug']);
   grunt.registerTask('build-firefox-prod', ['clean:build', 'pre-dist', 'copy:config_default', 'bundle-firefox', 'clean:debug_data', 'shell:build_quickaccess_prod', 'shell:build_firefox_prod']);
 
   grunt.registerTask('build-chrome', ['build-chrome-debug', 'build-chrome-prod']);
-  grunt.registerTask('build-chrome-debug', ['clean:build', 'pre-dist', 'copy:config_debug', 'bundle-chrome', 'bundle-test', 'shell:build_quickaccess_debug', 'shell:build_chrome_debug']);
+  grunt.registerTask('build-chrome-debug', ['clean:build', 'pre-dist', 'copy:config_debug', 'bundle-chrome', 'shell:build_quickaccess_debug', 'shell:build_chrome_debug']);
   grunt.registerTask('build-chrome-prod', ['clean:build', 'pre-dist', 'copy:config_default', 'bundle-chrome', 'clean:debug_data', 'shell:build_quickaccess_prod', 'shell:build_chrome_prod']);
 
-  grunt.registerTask('test', ['shell:test_quickaccess']);
+  grunt.registerTask('test', ['shell:test']);
 
   /**
    * Main grunt tasks configuration
@@ -100,10 +97,6 @@ module.exports = function(grunt) {
       app: {
         src: [path.src_addon + 'index.js'],
         dest: path.build + 'index.min.js'
-      },
-      test: {
-        src: [path.test + 'test.js'],
-        dest: path.build + 'test.min.js'
       }
     },
 
@@ -146,12 +139,6 @@ module.exports = function(grunt) {
       background_page: {
         files: [
           {expand: true, cwd: path.src_addon, src: 'index.html', dest: path.build }
-        ]
-      },
-      test: {
-        files: [
-          {expand: true, cwd: path.test, src: 'test.html', dest: path.build},
-          {expand: true, cwd: path.node_modules + 'mocha', src: ['mocha.js', 'mocha.css'], dest: path.build_vendors},
         ]
       },
       data: {
@@ -305,9 +292,9 @@ module.exports = function(grunt) {
       watch_quickaccess_debug: {
         command: "./node_modules/.bin/webpack --watch --config webpack.dev.config.js"
       },
-      test_quickaccess: {
+      test: {
         stdout: true,
-        command: "jest --config ./src/all/data/js/quickaccess/popup/.jest.config.json --no-cache ./src/all/data/js/quickaccess/popup --maxWorkers=4"
+        command: "jest --config .jest.config.json --no-cache ./src/all/ --maxWorkers=4"
       },
 
       /**
@@ -394,11 +381,6 @@ module.exports = function(grunt) {
       vendors: {
         files: [path.src + 'background_page/vendors.js', path.src + 'background_page/vendors/**/*.js', path.src + 'background_page/sdk/storage.js'],
         tasks: ['browserify:vendors'],
-        options: {spawn: false}
-      },
-      test: {
-        files: [path.test + '**/*.js', path.src + '**/**.js'],
-        tasks: ['browserify:test'],
         options: {spawn: false}
       }
     }